requirements.txt
black==19.10b0
Django==2.2.17
django-dynamic-filenames==1.1.4
django-environ==0.4.5
django-extensions==3.1.0
django-model-utils==4.0.0
django-rest-framework==0.1.0
django-stdimage==5.1.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
env==0.1.0
marshmallow==3.7.1
nodeenv==1.4.0
Pillow==7.2.0
pre-commit==2.6.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytest==6.1.0
pytest-cov==2.10.1
pytest-django==4.1.0
requests==2.24.0
